Use hidden status on creation
