+++ packages/common/src/services/decision/submitProposal.ts
@@ -1,11 +1,6 @@
 import { getTipTapClient } from '@op/collab';
 import { db, eq } from '@op/db/client';
-import {
-  type ProcessInstance,
-  ProposalStatus,
-  Visibility,
-  proposals,
-} from '@op/db/schema';
+import { type ProcessInstance, ProposalStatus, proposals } from '@op/db/schema';
 import { assertAccess, permission } from 'access-zones';
 
 import { CommonError, NotFoundError, ValidationError } from '../../utils';
@@ -90,12 +85,6 @@ export const submitProposal = async ({
     );
   }
 
-  const currentPhase = instanceData.phases.find(
-    (p) => p.phaseId === currentPhaseId,
-  );
-  const defaultHidden =
-    currentPhase?.rules?.proposals?.defaults?.hidden === true;
-
   // Validate proposal data against the proposal template schema
   const proposalTemplate = await resolveProposalTemplate(
     instanceData,
@@ -144,7 +133,6 @@ export const submitProposal = async ({
       .update(proposals)
       .set({
         status: ProposalStatus.SUBMITTED,
-        ...(defaultHidden ? { visibility: Visibility.HIDDEN } : {}),
         ...(proposalDataUpdate ? { proposalData: proposalDataUpdate } : {}),
       })
       .where(eq(proposals.id, data.proposalId))
